+++ packages/bbui/src/Button/Button.svelte
@@ -1,7 +1,10 @@
 <script>
   import "@spectrum-css/button/dist/index-vars.css"
   import Tooltip from "../Tooltip/Tooltip.svelte"
+  import { createEventDispatcher } from "svelte"
 
+  export let type
+  export let preventDefault = true
   export let disabled = false
   export let size = "M"
   export let cta = false
@@ -17,10 +20,18 @@
   export let id
 
   let showTooltip = false
+  const dispatch = createEventDispatcher()
+
+  const handleClick = e => {
+    if (preventDefault) e.preventDefault()
+
+    dispatch("click", e)
+  }
 </script>
 
 <button
   {id}
+  {type}
   class:spectrum-Button--cta={cta}
   class:spectrum-Button--primary={primary}
   class:spectrum-Button--secondary={secondary}
@@ -32,7 +43,7 @@
   class:disabled
   class="spectrum-Button spectrum-Button--size{size.toUpperCase()}"
   {disabled}
-  on:click|preventDefault
+  on:click={handleClick}
   on:mouseover={() => (showTooltip = true)}
   on:focus={() => (showTooltip = true)}
   on:mouseleave={() => (showTooltip = false)}
@@ -73,6 +84,7 @@
   button {
     position: relative;
   }
+
   .spectrum-Button-label {
     white-space: nowrap;
     overflow: hidden;
