+++ packages/frontend-core/src/components/index.js
@@ -1,3 +1,4 @@
 export { default as SplitPage } from "./SplitPage.svelte"
 export { default as TestimonialPage } from "./TestimonialPage.svelte"
+export { default as Testimonial } from "./Testimonial.svelte"
 export { Grid } from "./grid"
